Fix the status bar chip not able to click when HUN is showing issue
Context: the StatusBar (StatusBarWindow) and the Notification Shade
(SceneWindowRootView) are on different root Windows, and
SceneWindowRootView is z-indexed on top of StatusBarWindow. There's no
way to pass input through one window to the window below it using the
Android touch dispatching without spy windows or legacy monitors. The
only two ways to dispatch the touch event to StatusBarWindow are either:
proactively pass the event from the top Window, or claim a region of
top Window as not touchable.

Before flexi-glass, the NotificationShadeWindowView (the legacy root
Window, now is a child of SceneWindowRootView) proactively sends the
touch to the StatusBarViewController through an InteractionEventHandler
when it detacts the touch happened within the bound of status bar.

In flexi-glass, the NotificationShadeWindowView is no longer the top
level Window View, or the container of the NotificationStack, so the
touches are not dispatched by it before the Notification (HUN) gets the
touch. Thus the InteractionEventHandler is not forwarding the touch to
StatusBar.

This change modifies the logic of getTouchableRegion for the
HeadsUpManagerImpl to avoid the region above the showing HUN being
touchable for the Notification when a HUN is showing, to make the Status
Bar below it to have a chance to get the click.

The existing logic of stretchToTop marks the status bar region as
clickable when the top of HUN less equals to the status bar's bottom,
this change avoided using 0 as the top of HUN touchable region.

Fix: 452645744
Flag: com.android.systemui.scene_container
Test: Try to click the status bar chip when a HUN is showing
Change-Id: I1d0a5c498652cdb0637875e408ff302d4ddb49a0