Run dismiss action when keyguard dismissed below shade
This ensures that, if a dismissAction is added while the shade is open
on the lockscreen, and the device is already unlocked (either has swipe
security, or was unlocked by the face while the shade was open), the
action gets executed without requiring the shade to be closed first.

isOnShadeWhileUnlocked was meant to handle similar cases, but it was
emitting before the dismissAction was being set in this case, so it only
worked when a bouncer was being shown. I renamed it to better reflect
what it's meant to do, and I'm collecting it separately instead of
merging with finishedTransitionToGone because the logs were misleading
previously.

Fix: 466055286
Test: KeyguardDismissActionInteractorTest
Test: manually tested with different types of authentication (PIN, face,
fingerprint and multiple thereof)
Test: manually tested that guts don't open if you fail to authenticate
Flag: com.android.systemui.scene_container

Change-Id: I3d8a7d96599d9d29646ae3a9e2075a213568cb56