Move bubble to fullscreen if the task is no longer valid to bubble.
This change does some refactoring and consolidates the error toast and moving to fullscreen behavior to BubbleBarExpandedView's onTaskInfoChanged instead of having it partially in BubblesUnfoldListener. This is because whether a Foldable device is large screen or a small screen (unfolded or folded), a determining factor for if a task is valid to bubble, is only available in onTaskInfoChanged AFTER onFoldStateChanged is called, so handling behvaior should just be in onTaskInfoChanged to actually work.

Demo: https://screencast.googleplex.com/cast/NjI1MTQ2MjMxMjA2NzA3MnxjZDM1NmUzYS1mNg
Fix: 465967201
Test: atest WMShellMultivalentTestsOnDevice:BubbleBarExpandedViewTest
Test: atest WMShellUnitTests:DefaultBubblePolicyHelperTest
Flag: com.android.wm.shell.enable_create_any_bubble
Change-Id: I47269f6cb5f5c8a1b0ed2161b6cca3110699a489