Go to lockscreen after dream timeout
This change adds some logic in the DreamStartable to listen for dreaming
state if the device has an unsecured authentication method. When dream
starts it sends a delayed action to add lockscreen scene to the bottom
of the scene stack. This makes sure that when dream is awoken we go to
the lockscreen instead of home screen.

Test: atest DreamStartableTest
Test: set authentication method to "swipe", enable dreams, and wait for
      device timeout on the home screen. Verified that waking up the
      dream immediately goes back to the home screen, and waking it up
      after 5s goes to the lockscreen
Bug: 453749672
Flag: com.android.systemui.scene_container
Change-Id: Ifc84f2dbefe7226f6a9bdaef794508d54faa89df