Add class to provide fingerprinting of USB devices
For several use-cases that require persisting decisions on connected USB
devices, we need a way to uniquely identify USB devices.
UsbDeviceFingerprint uses a combination of a hash of the USB descriptor
and, optionally, the hash of all descriptors from root hub to the USB
device in order to uniquely identify the device. It also provides
different hash quality levels to reflect the uniqueness of the attached
device (i.e. USB devices with serial numbers are considered the most
unique).

Bug: 467507423
Test: Build and unit tests from subsequent commits
Flag: android.hardware.usb.flags.enable_persistent_device_permissions
Change-Id: I6fda96b625d30ddde2914a2131fcf3b8629dd4e3