Fix back animation background stuck on screen
The bug was reproducible by the following steps:
1. Starting a cross task back animation
2. Cancelling the back gesture
3. Immediately tap in the back gesture area (while the cancellation
   animation has not yet settled).

The problem was that this sequence of events interrupted the
cancellation animation but did not ensure to clean it up eventually if
`onThresholdCrossed()` was never called for the second user interaction.

This fix ensures that the animation (including background layer) is
properly cleaned up by going through the regular
startPostCommitAnimation() flow in this special case.

Bug: 462253806
Flag: EXEMPT BUGFIX
Test: Manual, i.e. verified bug is not reproducible with the fix
Change-Id: I5905fe3a11234c15a5ac321715d6931481332a5d