Implement TV PiP2 bounds change animation
Introduce new TvPipScheduler class to schedule bounds change animations,
which prompts the Transition framework to call TvPipTransition to run
the animation. Implement bounds change animation in TvPipTransition. Update the PipResizeAnimator to be able to sync the menu. Add a special case to adjust the menu bounds during edu text drawer close when the pip window is in the top corners since the bounds change animation is not triggered when they are positioned there.
This addition does not affect the functionality of PiP1.

Test: m
Flag: com.android.wm.shell.enable_pip2_on_tv
Bug: 403347147
Bug: 463390460
Change-Id: I1fb63071603f3b53b6647359d47c5c712d745a93