Don't remove `BoundsBeforeSnapOrMaximize` on drag move
`BoundsBeforeSnapOrMaximize` stores the bounds before a task enters a fixed size state like snap left/right or maximise so they can be restored when the task leaves this state. When the user drags out of one of these states we restore the bounds onDragMove so the task visually returns to the previous size while the user is positioning it. However, we should not remove the stored previous bounds while moving in case the task in drag into another fixed size state to ensure the original previous bounds are persisted. Instead the previous bounds will be cleared onDragEnd where we can establish if they are still needed or not.

Flag: com.android.window.flags.enable_bounds_restoring_on_drag_exit
Test: manually dragging between snap/ maximise and toggle maximise
Test: atest WMShellUnitTests:MultiDisplayVeiledResizeTaskPositionerTest
Fixes: 465992835
Change-Id: I25730aec1b8a3e3ee8aec1d930113d30312457cf