Show selected subtype's icon on IME chip if the IME provided an icon.
If there is a selected subtype and the IME has provided an icon for this
subtype, show it on the status bar IME chip. Otherwise, fall back to a
default keyboard icon.

We keep track of the icon resource id and package name in
InputMethodModel.Subtype, and use these to load the icon from
ImeIndicatorChipViewModel if needed.

In a follow-up, we will also make it possible to show the subtype short
label on the IME chip (so the priority is subtype icon > subtype short
label > default icon).

Bug: 458557858
Flag: com.android.systemui.status_bar_ime_chip
Test: atest SystemUITests:ImeIndicatorChipViewModelTest &&
  atest SystemUITests:InputMethodRepositoryTest
Change-Id: I3ccb0db5f6bdb5dec55994f43faa63d085c2660f