Show short label on IME chip if selected subtype doesn't have an icon.
The short label of an IME subtype is an optional non-localized string an
IME can specify to represent a subtype in compact UI elements in place
of an icon. See ag/36713227.

If there is a selected subtype and the IME has provided a short label
but not an icon for this subtype, show the short label on the IME chip.
Otherwise, fall back to the default keyboard icon.

Bug: 458557858
Flag: com.android.systemui.status_bar_ime_chip
Flag: android.view.inputmethod.ime_subtype_short_label
Test: atest SystemUITests:ImeIndicatorChipViewModelTest &&
  atest SystemUITests:InputMethodRepositoryTest
Change-Id: Ia44d359fbb949d8fe7695ddb5c368f68ddc0eefc