Do not preserve global focus if task is restored to inactive desk.
When restoring a display, we move the globally focused task to the front to preserve global focus. However, when the task in question was moved to an inactive desk, the inactive desk is wrongly moved to the front with it.

To fix this, confirm the globally focused task is not moved to an inactive desk before reordering it.

Bug: 459828420
Test: Manual
Flag: com.android.window.flags.enable_display_reconnect_interaction
Change-Id: If65436e135951d10e7b45c28c33ff6bb3385a6c1