Camera: Update cached surface sizes after 'updateOutputConfigurations'
Both the internal and client provided output configuration configured
sizes must be updated after 'updateOutputConfigurations' in order
to avoid potential discrepancies between the already configured
and the incoming new surface. The delta is possible because some
surfaces such as SurfaceTexture/SurfaceView etc. may have flexible
surface size and and the camera device can potentially override them.
As a result, we need the client side to be updated as well, so that
future OutputConfiguration updates do not fail compatibility checks.

Flag: NONE bugfix
Test: Camera CTS
Bug: 466630847
Change-Id: Ic91be5736cbba23393ab67f4c24784111457d56b