Usb: Implement persistent permissions for devices
Permissions for UsbDevice are currently temporary and disappear when the
device is disconnected. To persist the permission, we get a fingerprint
of the device and save it to an XML file.

This change refactors an earlier implementation to use
a UsbDeviceFingerprint class, which provides serialization and
deserialization. The new implementation provides persistence for
UsbDevice but not for UsbAccessory (intentionally as the identifying
information for accessories was found to be lacking for uniqueness and
is often empty in many implementations).

Also add basic tests for permission management testing the hasPermission and
grantPermission workflow and serialization.

Bug: 389062495
Test: atest UsbTests
Flag: android.hardware.usb.flags.enable_persistent_usb_device_permissions
Change-Id: If5cce0bd1cf98efe9a4704f419f40ad308a6c745