Replace what appear to be block tag uses of {@see ...} with @see ...
According to the javadoc specification `@see` is a block tag but
unfortunately `doclava` also allows it to be specified as an inline
tag, e.g. `{@see ...}`. In that case it behaves as if it was specified
as a block tag immediately before the first block tag when used in the
main description or before the current block tag when used in a block
tag description.

This change replaces all uses of `{@see ...}` that appear to be used as
a block tag, i.e. on its own line, no leading or trailing text with
block tags. Moving them to their standard location if they are in the
middle of the main description.

Flag: EXEMPT DOCS_ONLY
Bug: 469362503
Test: m offline-system-sdk-referenceonly-docs
      # Check before and after and make sure there are no significant
      # changes. The only changes are whitespace.
Change-Id: Ie21db9eebfa16ff4737ecbff6079f44af424eb14