Add a new permission, android.permission.ATTRIBUTE_WORK_TO_OTHER_APPS
This is a `privileged|signature` permission that grants the app the
ability to attribute work (such as NPU usage) to other apps. This is
privileged because it could be done without the presence of an
`AttributionSource` or other securely-obtained source of work, so the
app has the ability to lie.

Bug: 416283422
Bug: 469432860
Flag: com.android.npumanager.npumanager_enabled
Test: build
Change-Id: I419ed3f9e360c6aa16d136b16c41ccb4e10fdce4