top-row: Ensure correct display to show dialogs
Current, `dialogFactory.create` is to create dialog based on a default
application context. Sometimes, the diaplayId requested isn't the
default display. For this kind of case, we are trying to generate a
new context based on the diaplayId requested to create dialogs.

Bug: 458458326
Flag: android.view.accessibility.enable_a11y_top_row_shortcut
Test: atest ShortcutChooserDialogInteractorTest

Change-Id: I919dd24eecfd0bd22b58483a070e8737d090147c