Fix IME Switcher Menu overflow
When there are enough items in the list to start scrolling, and the
screen is small enough (e.g. landscape), the whole menu becomes
scrollable. However, this currently leads to hiding the settings button.

This fixes the issue by wrapping the LazyColumn in another Column with
weight 1.

Flag: android.view.inputmethod.ime_switcher_menu_systemui
Bug: 460776726
Test: atest InputMethodManagerTest#testInputMethodPickerShownItems
  InputMethodManagerTest#testInputMethodPickerSwitchIme
  InputMethodServiceTest#testImeSwitchButtonLongClick
  InputMethodServiceTest#testImeSwitcherMenu_openLanguageSettings
  InputMethodServiceTest#testImeSwitcherMenu_noLanguageSettingsWhenScreenLocked
  InputMethodServiceTest#testImeSwitcherMenu_noLanguageSettingsWhenDeviceNotProvisioned
Change-Id: Iad0655317786d77f42a8c31ee2bd005f34807324