[flexiglass] Update "fake AOD" touch handling
"Fake AOD" occurs when SysUI is transitioning to AOD/asleep,
but the display is still on, so touches still go to SysUI
(as opposed to "real AOD" which is when the device display state
is DOZE and touches do NOT go to SysUI).

We don't want to allow swipes to wake the device
in "fake AOD" to avoid falsing - only true taps should
wake the device (ie: touchSlop doesn't pass threshod).

Flag: com.android.systemui.scene_container
Test: Press power button to transition to AOD. In the first
couple seconds, tap the device. Observe device wakes up.
Test: After a reboot with UDFPS enrolled: press power button
to transition to AOD. In the first couple seconds, long press
the lock icon on AOD. Observe bouncer appears.
Fixes: 470135608

Change-Id: Ic4de7fd647d533f6ec3a12e38f7cd1049d9d598f