Collect data inside restoreDisplay coroutine.
In restoreDisplay, we gather needed variables before launching the main coroutine. This leaves a possibility for the values of those variables to be outdated when the coroutine ends up running. This CL fixes this by gathering the needed data inside the coroutine.

Bug: 469492330
Test: Manual
Flag: com.android.window.flags.enable_display_reconnect_interaction
Change-Id: Ia38a1f22d053aa59b9441687c90d60a33d5fc2e9